+++ lib/src/widgets/email_text_field.dart
@@ -31,9 +31,11 @@ class EmailTextField extends StatelessWidget {
               validator: emailElement!.validator,
               readOnly: emailElement!.readOnly,
               keyboardType: Constants.getInput(emailElement!.typeInput),
-              style: emailElement!.decorationElement.style,
+              style: emailElement!.decorationElement?.style,
               focusNode: currentFocus,
-              textInputAction: nextFocus == null ? TextInputAction.done : TextInputAction.next,
+              textInputAction: nextFocus == null
+                  ? TextInputAction.done
+                  : TextInputAction.next,
               onFieldSubmitted: (v) {
                 Constants.fieldFocusChange(context, currentFocus, nextFocus);
               },
@@ -50,9 +52,10 @@ class EmailTextField extends StatelessWidget {
       validator: emailElement!.validator,
       readOnly: emailElement!.readOnly,
       keyboardType: Constants.getInput(emailElement!.typeInput),
-      style: emailElement!.decorationElement.style,
+      style: emailElement!.decorationElement?.style,
       focusNode: currentFocus,
-      textInputAction: nextFocus == null ? TextInputAction.done : TextInputAction.next,
+      textInputAction:
+          nextFocus == null ? TextInputAction.done : TextInputAction.next,
       onFieldSubmitted: (v) {
         Constants.fieldFocusChange(context, currentFocus, nextFocus);
       },