+++ lib/src/widgets/phone_text_field.dart
@@ -40,7 +40,8 @@ class _CallingCountry {
       : this.fullName = map["name"],
         this.code3Alpha = map["alpha3Code"],
         this.code2Alpha = map["alpha2Code"],
-        this.callingCode = (map["callingCodes"] as List).map((c) => "+${c.trim()}").toList();
+        this.callingCode =
+            (map["callingCodes"] as List).map((c) => "+${c.trim()}").toList();
 
   @override
   int get hashCode {
@@ -93,7 +94,8 @@ class _PhoneTextFieldState extends State<PhoneTextField> {
 
   void changeTextListener() {
     final callingCode = countryNotifier.value?.callingCode.first ?? "";
-    widget.controller.text = callingCode.trim() + textEditingController.text.trim();
+    widget.controller.text =
+        callingCode.trim() + textEditingController.text.trim();
   }
 
   @override
@@ -142,9 +144,12 @@ class _PhoneTextFieldState extends State<PhoneTextField> {
             readOnly: widget.element.readOnly,
             style: TextStyle(color: Colors.black),
             focusNode: widget.currentFocus,
-            textInputAction: widget.nextFocus == null ? TextInputAction.done : TextInputAction.next,
+            textInputAction: widget.nextFocus == null
+                ? TextInputAction.done
+                : TextInputAction.next,
             onFieldSubmitted: (v) {
-              Constants.fieldFocusChange(context, widget.currentFocus, widget.nextFocus);
+              Constants.fieldFocusChange(
+                  context, widget.currentFocus, widget.nextFocus);
             },
             decoration: Constants.setInputBorder(
               context,
@@ -170,9 +175,12 @@ class _PhoneTextFieldState extends State<PhoneTextField> {
       readOnly: widget.element.readOnly,
       style: TextStyle(color: Colors.black),
       focusNode: widget.currentFocus,
-      textInputAction: widget.nextFocus == null ? TextInputAction.done : TextInputAction.next,
+      textInputAction: widget.nextFocus == null
+          ? TextInputAction.done
+          : TextInputAction.next,
       onFieldSubmitted: (v) {
-        Constants.fieldFocusChange(context, widget.currentFocus, widget.nextFocus);
+        Constants.fieldFocusChange(
+            context, widget.currentFocus, widget.nextFocus);
       },
       decoration: Constants.setInputBorder(
         context,
@@ -193,8 +201,10 @@ class _PhoneTextFieldState extends State<PhoneTextField> {
                     iconFlag,
                     if (prefixWidget != null) ...[
                       Padding(
-                        padding: EdgeInsets.only(left: 5,top: 6),
-                        child: Center(child: prefixWidget,),
+                        padding: EdgeInsets.only(left: 5, top: 6),
+                        child: Center(
+                          child: prefixWidget,
+                        ),
                       )
                     ],
                   ],
@@ -243,7 +253,10 @@ class _PrefixPhoneNumberState extends State<PrefixPhoneNumber> {
         (data) => _CallingCountry.fromJson(data),
       ).then((list) {
         final index = widget.prefix.isNotEmpty
-            ? list.map((e) => e.callingCode.first).toList().indexOf(widget.prefix)
+            ? list
+                .map((e) => e.callingCode.first)
+                .toList()
+                .indexOf(widget.prefix)
             : 0;
 
         widget.countryNotifier.value = list[index].copy(
@@ -262,7 +275,9 @@ class _PrefixPhoneNumberState extends State<PrefixPhoneNumber> {
             context: context,
             builder: (ctx) {
               return _CallingCodeModalPopups(
-                initCallingCode: widget.countryNotifier.value?.callingCode.first ?? widget.prefix,
+                initCallingCode:
+                    widget.countryNotifier.value?.callingCode.first ??
+                        widget.prefix,
                 title: widget.title,
                 selectCallingCodeFunction: (callingCode) async {
                   widget.countryNotifier.value = callingCode;
@@ -281,7 +296,9 @@ class _PrefixPhoneNumberState extends State<PrefixPhoneNumber> {
             )),
             builder: (ctx) {
               return _CallingCodeModalPopups(
-                initCallingCode: widget.countryNotifier.value?.callingCode.first ?? widget.prefix,
+                initCallingCode:
+                    widget.countryNotifier.value?.callingCode.first ??
+                        widget.prefix,
                 title: widget.title,
                 selectCallingCodeFunction: (callingCode) async {
                   widget.countryNotifier.value = callingCode;
@@ -362,12 +379,15 @@ class _CallingCodeModalPopups extends StatelessWidget {
                     )
                     .expand((element) => element)
                     .toList();
-                list.removeWhere((element) => element.callingCode.first.length == 1);
+                list.removeWhere(
+                    (element) => element.callingCode.first.length == 1);
                 return ListView.builder(
                   controller: ScrollController(
-                    initialScrollOffset:
-                        list.map((e) => e.callingCode.first).toList().indexOf(initCallingCode) *
-                            55.5,
+                    initialScrollOffset: list
+                            .map((e) => e.callingCode.first)
+                            .toList()
+                            .indexOf(initCallingCode) *
+                        55.5,
                   ),
                   itemExtent: 56,
                   itemBuilder: (ctx, index) {
@@ -388,9 +408,10 @@ class _CallingCodeModalPopups extends StatelessWidget {
                         overflow: TextOverflow.ellipsis,
                         maxLines: 2,
                         style: TextStyle(
-                          color: countryInfo.callingCode.first == initCallingCode
-                              ? Theme.of(context).primaryColor
-                              : Theme.of(context).colorScheme.onSurface,
+                          color:
+                              countryInfo.callingCode.first == initCallingCode
+                                  ? Theme.of(context).primaryColor
+                                  : Theme.of(context).colorScheme.onSurface,
                         ),
                       ),
                       trailing: Text(